Rename root package from whynot_web to whynot
