+++ docker-compose-dev.yml
@@ -6,6 +6,6 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    command: watchmedo auto-restart -d /usr/src/app -p "*.py" -D -R -- gunicorn -k gevent -b 0.0.0.0:15000 whynot_web.application:app
+    command: watchmedo auto-restart -d /usr/src/app -p "*.py" -D -R -- gunicorn -k gevent -b 0.0.0.0:15000 whynot.application:app
     environment:
       - WHYNOT_SETTINGS=/usr/src/app/dev_settings.py
