+++ docker-compose.yml
@@ -6,7 +6,7 @@ services:
     build:
       context: .
       dockerfile: Dockerfile
-    command: gunicorn -k gevent -b 0.0.0.0:15000 whynot_web.application:app
+    command: gunicorn -k gevent -b 0.0.0.0:15000 whynot.application:app
     environment:
       - WHYNOT_SETTINGS=/usr/src/app/prd_settings.py
     ports:
