+++ whynot/application.py
@@ -1,0 +1,3 @@
+from whynot.factory import create_app
+
+app = create_app()
