+++ whynot/frontend/templates/search/SearchPanel.html
