--- whynot_web/application.py
@@ -1,3 +1,0 @@
-from whynot_web.factory import create_app
-
-app = create_app()
